Run parser microbenchmarks in simulation only
